# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SleepStage = typeof SleepStage[keyof typeof SleepStage];


export const SleepStage = {
  REM: 'REM',
  NREM: 'NREM',
  Awake: 'Awake',
} as const;

export interface SensorReading {
  timestamp: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  spo2: number;
  /** @minimum 0 */
  heart_rate: number;
  temperature: number;
  movement: boolean;
  sleep_stage: SleepStage;
}

export interface IngestBatchRequest {
  /** Reuse across batch posts to append to the same session */
  sessionId?: string;
  /**
     * @minItems 1
     * @maxItems 1000
     */
  readings: SensorReading[];
}

export interface IngestBatchResponse {
  sessionId: string;
  accepted: number;
  totalInSession: number;
}

export interface IngestSessionSummary {
  sessionId: string;
  readingCount: number;
  firstTimestamp?: string | null;
  lastTimestamp?: string | null;
}

export interface IngestSessionListResponse {
  sessions: IngestSessionSummary[];
}

export interface SessionReadingsResponse {
  sessionId: string;
  count: number;
  readings: SensorReading[];
}

export interface ErrorResponse {
  error: string;
}