# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ingest
    description: Sensor reading batch ingest (ML / device)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ingest/batch:
    post:
      operationId: ingestReadingsBatch
      tags: [ingest]
      summary: Ingest a batch of sensor readings
      description: |
        Accepts up to 1000 readings per request. Reuse the same sessionId across
        batches to build a full session. Omit sessionId to auto-create one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBatchRequest"
      responses:
        "201":
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBatchResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ingest/sessions:
    get:
      operationId: listIngestSessions
      tags: [ingest]
      summary: List ingested sessions
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestSessionListResponse"
  /ingest/sessions/{sessionId}/readings:
    get:
      operationId: getSessionReadings
      tags: [ingest]
      summary: Get all readings for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Readings for session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionReadingsResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SleepStage:
      type: string
      enum: [REM, NREM, Awake]
    SensorReading:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        spo2:
          type: number
          minimum: 0
          maximum: 100
        heart_rate:
          type: integer
          minimum: 0
        temperature:
          type: number
        movement:
          type: boolean
        sleep_stage:
          $ref: "#/components/schemas/SleepStage"
      required:
        - timestamp
        - spo2
        - heart_rate
        - temperature
        - movement
        - sleep_stage
    IngestBatchRequest:
      type: object
      properties:
        sessionId:
          type: string
          description: Reuse across batch posts to append to the same session
        readings:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/SensorReading"
      required:
        - readings
    IngestBatchResponse:
      type: object
      properties:
        sessionId:
          type: string
        accepted:
          type: integer
        totalInSession:
          type: integer
      required:
        - sessionId
        - accepted
        - totalInSession
    IngestSessionSummary:
      type: object
      properties:
        sessionId:
          type: string
        readingCount:
          type: integer
        firstTimestamp:
          type: string
          format: date-time
          nullable: true
        lastTimestamp:
          type: string
          format: date-time
          nullable: true
      required:
        - sessionId
        - readingCount
    IngestSessionListResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/IngestSessionSummary"
      required:
        - sessions
    SessionReadingsResponse:
      type: object
      properties:
        sessionId:
          type: string
        count:
          type: integer
        readings:
          type: array
          items:
            $ref: "#/components/schemas/SensorReading"
      required:
        - sessionId
        - count
        - readings
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error