# lib/api-zod/src/generated/types/ingestBatchRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SensorReading } from './sensorReading';

export interface IngestBatchRequest {
  /** Reuse across batch posts to append to the same session */
  sessionId?: string;
  /**
     * @minItems 1
     * @maxItems 1000
     */
  readings: SensorReading[];
}