# lib/api-zod/src/generated/types/ingestSessionListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IngestSessionSummary } from './ingestSessionSummary';

export interface IngestSessionListResponse {
  sessions: IngestSessionSummary[];
}