# lib/api-zod/src/generated/types/ingestSessionSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IngestSessionSummary {
  sessionId: string;
  readingCount: number;
  firstTimestamp?: Date | null;
  lastTimestamp?: Date | null;
}