# lib/api-zod/src/generated/types/sensorReading.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SleepStage } from './sleepStage';

export interface SensorReading {
  timestamp: Date;
  /**
     * @minimum 0
     * @maximum 100
     */
  spo2: number;
  /** @minimum 0 */
  heart_rate: number;
  temperature: number;
  movement: boolean;
  sleep_stage: SleepStage;
}