# lib/api-zod/src/generated/types/sessionReadingsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SensorReading } from './sensorReading';

export interface SessionReadingsResponse {
  sessionId: string;
  count: number;
  readings: SensorReading[];
}