# lib/api-zod/src/generated/types/sleepStage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type SleepStage = typeof SleepStage[keyof typeof SleepStage];


export const SleepStage = {
  REM: 'REM',
  NREM: 'NREM',
  Awake: 'Awake',
} as const;